Use a more recent kernel
This uses kernel 5.11.16 from https://github.com/tobetter/linux
with a custom config pulled from
https://docs.linuxfactory.or.kr/install/odroidhc4/image.html

It does not boot with petitboot (in the HC4 board flash), you need to
either hold the boot switch button while booting or disable petitboot
with these instructions:
https://forum.odroid.com/viewtopic.php?p=318658#p318658
(there are instructions for restoring it here:
https://forum.odroid.com/viewtopic.php?f=207&t=42217)

This splits the uboot module into two, one to build the uboot binary,
and another to handle generating the boot scripts.

Everything goes in the ext4 partition and the vfat partition is now
unused. The boot.scr file refers to all the files it needs directly from
the nix store.